+++ build_config/lib/build_config.dart
@@ -2,4 +2,15 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/build_config.dart';
+export 'src/build_config.dart'
+    show
+        AutoApply,
+        BuildConfig,
+        BuilderDefinition,
+        BuildTarget,
+        BuildTo,
+        InputSet,
+        TargetBuilderConfig,
+        TargetBuilderConfigDefaults;
+export 'src/key_normalization.dart'
+    show normalizeBuilderKeyUsage, normalizeTargetKeyUsage;