+++ build_runner/bin/create_merged_dir.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/args.dart';
@@ -52,8 +51,7 @@ Future main(List<String> args) async {
 
   await logTimedAsync(logger, 'Loading asset graph at ${assetGraphFile.path}',
       () async {
-    assetGraph = new AssetGraph.deserialize(
-        JSON.decode(await assetGraphFile.readAsString()) as Map);
+    assetGraph = new AssetGraph.deserialize(await assetGraphFile.readAsBytes());
   });
 
   packageGraph = new PackageGraph.forThisPackage();