+++ build_runner/lib/src/build_script_generate/build_script_generate.dart
@@ -18,11 +18,13 @@ import 'builder_ordering.dart';
 
 const scriptLocation = '$entryPointDir/build.dart';
 
-Future<String> generateBuildScript() => logTimedAsync(
-    new Logger('Entrypoint'), 'Generating build script', _generateBuildScript);
+Future<String> generateBuildScript([String configKey]) => logTimedAsync(
+    new Logger('Entrypoint'),
+    'Generating build script',
+    () => _generateBuildScript(configKey));
 
-Future<String> _generateBuildScript() async {
-  final builders = await _findBuilderApplications();
+Future<String> _generateBuildScript(String configKey) async {
+  final builders = await _findBuilderApplications(configKey);
   final library = new Library((b) => b.body.addAll(
       [literalList(builders).assignFinal('_builders').statement, _main()]));
   final emitter = new DartEmitter(new Allocator.simplePrefixing());
@@ -34,13 +36,14 @@ Future<String> _generateBuildScript() async {
 ///
 /// Adds `apply` expressions based on the BuildefDefinitions from any package
 /// which has a `build.yaml`.
-Future<Iterable<Expression>> _findBuilderApplications() async {
+Future<Iterable<Expression>> _findBuilderApplications(String configKey) async {
   final builderApplications = <Expression>[];
   final packageGraph = new PackageGraph.forThisPackage();
   final orderedPackages = stronglyConnectedComponents<String, PackageNode>(
           [packageGraph.root], (node) => node.name, (node) => node.dependencies)
       .expand((c) => c);
-  final buildConfigOverrides = await findBuildConfigOverrides(packageGraph);
+  final buildConfigOverrides =
+      await findBuildConfigOverrides(packageGraph, configKey);
   Future<BuildConfig> _packageBuildConfig(PackageNode package) async {
     if (buildConfigOverrides.containsKey(package.name)) {
       return buildConfigOverrides[package.name];
@@ -61,14 +64,25 @@ Future<Iterable<Expression>> _findBuilderApplications() async {
 /// A method forwarding to `run`.
 Method _main() => new Method((b) => b
   ..name = 'main'
-  ..lambda = true
+  ..modifier = MethodModifier.async
   ..requiredParameters.add(new Parameter((b) => b
     ..name = 'args'
     ..type = new TypeReference((b) => b
       ..symbol = 'List'
       ..types.add(refer('String')))))
-  ..body = refer('run', 'package:build_runner/build_runner.dart')
-      .call([refer('args'), refer('_builders')]).code);
+  ..optionalParameters.add(new Parameter((b) => b
+    ..name = 'sendPort'
+    ..type = refer('SendPort', 'dart:isolate')))
+  ..body = new Block.of([
+    refer('run', 'package:build_runner/build_runner.dart')
+        .call([refer('args'), refer('_builders')])
+        .awaited
+        .assignVar('result')
+        .statement,
+    refer('sendPort')
+        .nullSafeProperty('send')
+        .call([refer('result')]).statement,
+  ]));
 
 /// An expression calling `apply` with appropriate setup for a Builder.
 Expression _applyBuilder(BuilderDefinition definition) =>