+++ e2e_example/test/dart2js_integration_test.dart
@@ -4,6 +4,7 @@
 
 @TestOn('vm')
 @Timeout(const Duration(minutes: 5))
+import 'dart:async';
 import 'dart:io';
 
 import 'package:test/test.dart';
@@ -11,25 +12,34 @@ import 'package:test/test.dart';
 import 'common/utils.dart';
 
 void main() {
-  var dart2JsBuildFile = new File('build.yaml.dart2js');
-  var actualBuildFile = new File('build.yaml');
+  group('Can run tests using dart2js', () {
+    test('via build.yaml config flag', () async {
+      await expectTestsPass(
+          usePrecompiled: true,
+          useManualScript: false,
+          args: ['--config=dart2js']);
+      await expectWasCompiledWithDart2Js();
+    }, onPlatform: {'windows': const Skip('flaky on windows')});
 
-  setUpAll(() async {
-    expect(await dart2JsBuildFile.exists(), isTrue);
-    expect(await actualBuildFile.exists(), isFalse);
-    await actualBuildFile.writeAsString(await dart2JsBuildFile.readAsString());
+    test('via --define flag', () async {
+      await expectTestsPass(
+          usePrecompiled: true,
+          useManualScript: false,
+          args: [
+            '--define',
+            'build_web_compilers|entrypoint=compiler=dart2js',
+            '--define',
+            'build_web_compilers|entrypoint=dart2js_args=["--checked"]',
+          ]);
+      await expectWasCompiledWithDart2Js();
+    }, onPlatform: {'windows': const Skip('flaky on windows')});
   });
+}
 
-  tearDownAll(() async {
-    await actualBuildFile.delete();
-  });
-
-  test('Can run tests using dart2js', () async {
-    await expectTestsPass(usePrecompiled: true, useManualScript: false);
-    var jsFile =
-        new File('.dart_tool/build/generated/e2e_example/web/main.dart.js');
-    expect(await jsFile.exists(), isTrue);
-    // sanity check that it was indeed compiled with dart2js
-    expect(await jsFile.readAsString(), contains('dart2js'));
-  }, onPlatform: {'windows': const Skip('flaky on windows')});
+Future<Null> expectWasCompiledWithDart2Js() async {
+  var jsFile =
+      new File('.dart_tool/build/generated/e2e_example/web/main.dart.js');
+  expect(await jsFile.exists(), isTrue);
+  // sanity check that it was indeed compiled with dart2js
+  expect(await jsFile.readAsString(), contains('dart2js'));
 }