+++ e2e_example/test/goldens/generated_build_script.dart
@@ -3,6 +3,7 @@ import 'package:provides_builder/builders.dart' as _i2;
 import 'package:build_test/builder.dart' as _i3;
 import 'package:build_config/build_config.dart' as _i4;
 import 'package:build_web_compilers/builders.dart' as _i5;
+import 'dart:isolate' as _i6;
 
 final _builders = [
   _i1.apply('provides_builder|some_not_applied_builder', [_i2.notApplied],
@@ -11,9 +12,8 @@ final _builders = [
   _i1.apply('provides_builder|some_builder', [_i2.someBuilder],
       _i1.toDependentsOf('provides_builder'),
       hideOutput: true),
-  _i1.apply(
-      'build_test|test_bootstrap', [_i3.testBootstrapBuilder], _i1.toRoot(),
-      isOptional: true,
+  _i1.apply('build_test|test_bootstrap',
+      [_i3.debugTestBuilder, _i3.testBootstrapBuilder], _i1.toRoot(),
       hideOutput: true,
       defaultGenerateFor: const _i4.InputSet(include: const ['test/**'])),
   _i1.apply(
@@ -31,6 +31,10 @@ final _builders = [
       _i1.toRoot(),
       hideOutput: true,
       defaultGenerateFor: const _i4.InputSet(
-          include: const ['web/**', 'test/**.browser_test.dart']))
+          include: const ['web/**', 'test/**_test.dart'],
+          exclude: const ['test/**.node_test.dart', 'test/**.vm_test.dart']))
 ];
-main(List<String> args) => _i1.run(args, _builders);
+main(List<String> args, [_i6.SendPort sendPort]) async {
+  var result = await _i1.run(args, _builders);
+  sendPort?.send(result);
+}