+++ build_web_compilers/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.1
+
+- All dart files under `test` are now compiled by default instead of only the
+  `_browser_test.dart` files (minus vm/node test bootstrap files). This means
+  the original tests can be debugged directly (prior to package:test
+  bootstrapping).
+- Updated to `package:build` version `0.12.0`.
+
 # 0.2.0
 
 ## New Features